Calculate a DMG's extra space by adding x percent
Line 148 adds an extra amount of space to the temporary read/write DMG, over and above what the storage requirements were calculated to be in the previous line. When creating a DMG involving very large files (several gigabytes total), the creation of the DMG may fail because hdiutil reports that the DMG runs out of space. By adding a percentage of extra space to be added to the DMG (I chose an additional 20%), rather than a fixed amount (originally, 20MB, and in my initial experiments, 100M), I was able to create DMGs up to 19GB in size, when previously, I was only able to create DMGs up to approximately 3GB before encountering the out-of-space errors.